Find Endmembers from a list of LIA points

Finds endmembers from a set of Lead Isotope Points using Principle Component analysis and the Geochron slope according to the two-stage model.

endmembers(x, col = NULL, tolerance = c(0.01, 0.01), clamp = c(Inf, Inf), ...)

Arguments

x

data.frame or matrix object containing Pb 206/204, 207/204, 208/204 isotope ratios.

col

Isotope column names containing Pb 206/204, 207/204, 208/204 isotope ratios. Names must contains the significatn numbers 6, 7 and 8.

tolerance

Vector of length two, with corespoitng group 1 and group 2 tolerance value for points considered to be intercepted. (Default c(0.01, 0.01))

clamp

Limit filter for points away from the principle component end based on Euclidean distance, (Default c(Inf, Inf))

...

Additional Parameters

Value

An object of class 'liaendmembers' as a list of 6

data

Isotope data

pca_ends

Endmembers maped using PCA

group1,group2

End member groups

mixing

Mixing Group

tolerance

Totlarance value

clamp

Clamping values

pca

List of PCA analysis

Examples

# Create object with class liaendmembers
data("tel_dor")
end_members <- endmembers(
        tel_dor,
        colnames(tel_dor),
        tolerance = c(0.01, 0.01),
        clamp = c(Inf, Inf)
)
#> PC2 or PC3 are not normally distributed. This may indicate that their variation may not be random noise.
# Print summary of the liaendmembers object
summary(end_members)
#> Summary of End memebers:
#> 
#> Tolarance: 0.01 0.01 
#> Clamp: Inf Inf 
#> 
#> PCA Endmembers
#>      pb64   pb74   pb84
#> 9  17.906 15.643 37.991
#> 10 19.034 15.732 39.382
#> 
#>        Group1 Group2 Mixing Total
#> Counts      3     12     17    32
#> - - - - - - - - - - - - - - - - - - 
#> Importance of components:
#>                           PC1     PC2     PC3
#> Standard deviation     0.4957 0.05492 0.01592
#> Proportion of Variance 0.9869 0.01211 0.00102
#> Cumulative Proportion  0.9869 0.99898 1.00000