Euclidean Distance for Pb Isotope rations to ore Sources

Calcutalte the culidian distanc of each isotope sample to a reference dataset, and gives the clossest regions to the groups. Mass-fractionation follows the procedure outlined in Albarede et.al (2024)

euc_dist(x, ox, ox_col, group, .n = 10, as.dist = FALSE, ...)

mf_dist(x, ox, ox_col, group, .n = 10, as.dist = FALSE, s = 0.001, ...)

endmember_dist(
  x,
  ox,
  ox_col,
  group,
  dist_type = "ed",
  .n = 10,
  as.dist = FALSE,
  s = 0.001,
  ...
)

Arguments

x

Matrix of Pb Isotopes with colums in the order of 206Pb/204Pb, 207Pb/204Pb,208Pb/204Pb or a `liaendmember` object.

ox

Dataframe of renfrence data contained the Pb Isotope rations

ox_col

Vector contained names of the Pb Isotope rations. Names much include digits 6, 7, 8 for isotope ratio identification.

group

Groups belonging to reach isotope ration vector

.n

Length of result output (Default = 10)

as.dist

Option to show distances or only a vector of closest groups.

...

Additional params

s

Mass-fractionation factor (Defualt = 0.001)

dist_type

Distance type to use, simple euclidiant ('ed') or mass-fractionation corrected ('mfd')

Value

Data frame or charecter vector

List of dataframe or charecter vector

References

Albarede, F., Davis, G., Blichert-Toft, J., Gentelli, L., Gitler, H., Pinto, M., & Telouk, P. (2024). A new algorithm for using Pb isotopes to determine the provenance of bullion in ancient Greek coinage. Journal of Archaeological Science, 163, 105919. https://doi.org/10.1016/j.jas.2023.105919